Split import SF course transaction out of loop
